Make sure that the ruby callback proc is marked when helium conn is
RB_GC_GUARD is not sufficent to prevent the callback proc from being
garbage collected. Instead use the 'mark' hook in Data_Wrap_struct to
also mark the callback that the helium connection references. Thus, as
long as the helium connection is reachable by the mark phase, the
callback proc will be as well.